LayoutTests/imported/w3c/web-platform-tests/css/css-values/progress-serialize-expected.txt


PASS 'progress(100px, 0px, 100px)' as a specified value should serialize as 'calc(1)'.
PASS 'scale(progress(100px, 0px, 100px))' as a specified value should serialize as 'scale(calc(1))'.
PASS 'progress(100px, 0px, 100px)' as a computed value should serialize as '1'.
PASS 'scale(progress(100px, 0px, 100px))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
PASS 'progress(10em, 0px, 10em)' as a specified value should serialize as 'progress(10em, 0px, 10em)'.
PASS 'scale(progress(10em, 0px, 10em))' as a specified value should serialize as 'scale(progress(10em, 0px, 10em))'.
PASS 'progress(10em, 0px, 10em)' as a computed value should serialize as '1'.
PASS 'scale(progress(10em, 0px, 10em))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
PASS 'progress(10em, 0px, 10rem)' as a specified value should serialize as 'progress(10em, 0px, 10rem)'.
PASS 'scale(progress(10em, 0px, 10rem))' as a specified value should serialize as 'scale(progress(10em, 0px, 10rem))'.
PASS 'progress(10em, 0px, 10rem)' as a computed value should serialize as '1'.
PASS 'scale(progress(10em, 0px, 10rem))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
PASS 'progress(100px, (10px - 10px), 100px)' as a specified value should serialize as 'calc(1)'.
PASS 'scale(progress(100px, (10px - 10px), 100px))' as a specified value should serialize as 'scale(calc(1))'.
PASS 'progress(100px, (10px - 10px), 100px)' as a computed value should serialize as '1'.
PASS 'scale(progress(100px, (10px - 10px), 100px))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
PASS 'progress(1%, (10% - 10%), 100%)' as a specified value should serialize as 'calc(0.01)'.
PASS 'scale(progress(1%, (10% - 10%), 100%))' as a specified value should serialize as 'scale(calc(0.01))'.
PASS 'progress(1%, (10% - 10%), 100%)' as a computed value should serialize as '0.01'.
PASS 'scale(progress(1%, (10% - 10%), 100%))' as a computed value should serialize as 'matrix(0.01, 0, 0, 0.01, 0, 0)'.
PASS 'calc(0.5 * progress(100px, 0px, 100px))' as a specified value should serialize as 'calc(0.5)'.
PASS 'scale(calc(0.5 * progress(100px, 0px, 100px)))' as a specified value should serialize as 'scale(calc(0.5))'.
PASS 'calc(0.5 * progress(100px, 0px, 100px))' as a computed value should serialize as '0.5'.
PASS 'scale(calc(0.5 * progress(100px, 0px, 100px)))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'.
PASS 'calc(0.5 * progress(200px, 0px, 100px))' as a specified value should serialize as 'calc(0.5)'.
PASS 'scale(calc(0.5 * progress(200px, 0px, 100px)))' as a specified value should serialize as 'scale(calc(0.5))'.
PASS 'calc(0.5 * progress(200px, 0px, 100px))' as a computed value should serialize as '0.5'.
PASS 'scale(calc(0.5 * progress(200px, 0px, 100px)))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'.
PASS 'calc(0.5 * progress(-100px, 0px, 100px))' as a specified value should serialize as 'calc(0)'.
PASS 'scale(calc(0.5 * progress(-100px, 0px, 100px)))' as a specified value should serialize as 'scale(calc(0))'.
PASS 'calc(0.5 * progress(-100px, 0px, 100px))' as a computed value should serialize as '0'.
PASS 'scale(calc(0.5 * progress(-100px, 0px, 100px)))' as a computed value should serialize as 'matrix(0, 0, 0, 0, 0, 0)'.
FAIL 'progress(2rad, 1rad, 1rad)' as a specified value should serialize as 'calc(0)'. assert_equals: 'progress(2rad, 1rad, 1rad)' and 'calc(0)' should serialize the same in specified values. expected "calc(0)" but got "calc(1)"
FAIL 'progress(1rad, 1rad, 1rad)' as a specified value should serialize as 'calc(0)'. assert_equals: 'progress(1rad, 1rad, 1rad)' and 'calc(0)' should serialize the same in specified values. expected "calc(0)" but got "calc(NaN)"
PASS 'progress(0rad, 1rad, 1rad)' as a specified value should serialize as 'calc(0)'.
PASS 'calc(50px * progress(100px, 0px, 100px))' as a specified value should serialize as 'calc(50px)'.
PASS 'calc(1px * progress(abs(10%), (10% - 10%), 100% / 10))' as a computed value should serialize as '1px'.
FAIL 'progress(no-clamp 100px, 0px, 100px)' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'progress(no-clamp 100px, 0px, 100px)' should be valid in opacity. got disallowed value ""
FAIL 'scale(progress(no-clamp 100px, 0px, 100px))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(progress(no-clamp 100px, 0px, 100px))' should be valid in transform. got disallowed value ""
PASS 'progress(no-clamp 100px, 0px, 100px)' as a computed value should serialize as '1'.
FAIL 'scale(progress(no-clamp 100px, 0px, 100px))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(progress(no-clamp 100px, 0px, 100px))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'progress(no-clamp 10em, 0px, 10em)' as a specified value should serialize as 'progress(no-clamp 10em, 0px, 10em)'. assert_not_equals: 'progress(no-clamp 10em, 0px, 10em)' should be valid in opacity. got disallowed value ""
FAIL 'scale(progress(no-clamp 10em, 0px, 10em))' as a specified value should serialize as 'scale(progress(no-clamp 10em, 0px, 10em))'. assert_not_equals: 'scale(progress(no-clamp 10em, 0px, 10em))' should be valid in transform. got disallowed value ""
PASS 'progress(no-clamp 10em, 0px, 10em)' as a computed value should serialize as '1'.
FAIL 'scale(progress(no-clamp 10em, 0px, 10em))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(progress(no-clamp 10em, 0px, 10em))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'progress(no-clamp 10em, 0px, 10rem)' as a specified value should serialize as 'progress(no-clamp 10em, 0px, 10rem)'. assert_not_equals: 'progress(no-clamp 10em, 0px, 10rem)' should be valid in opacity. got disallowed value ""
FAIL 'scale(progress(no-clamp 10em, 0px, 10rem))' as a specified value should serialize as 'scale(progress(no-clamp 10em, 0px, 10rem))'. assert_not_equals: 'scale(progress(no-clamp 10em, 0px, 10rem))' should be valid in transform. got disallowed value ""
PASS 'progress(no-clamp 10em, 0px, 10rem)' as a computed value should serialize as '1'.
FAIL 'scale(progress(no-clamp 10em, 0px, 10rem))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(progress(no-clamp 10em, 0px, 10rem))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'progress(no-clamp 100px, (10px - 10px), 100px)' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'progress(no-clamp 100px, (10px - 10px), 100px)' should be valid in opacity. got disallowed value ""
FAIL 'scale(progress(no-clamp 100px, (10px - 10px), 100px))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(progress(no-clamp 100px, (10px - 10px), 100px))' should be valid in transform. got disallowed value ""
PASS 'progress(no-clamp 100px, (10px - 10px), 100px)' as a computed value should serialize as '1'.
FAIL 'scale(progress(no-clamp 100px, (10px - 10px), 100px))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(progress(no-clamp 100px, (10px - 10px), 100px))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'progress(no-clamp 1%, (10% - 10%), 100%)' as a specified value should serialize as 'calc(0.01)'. assert_not_equals: 'progress(no-clamp 1%, (10% - 10%), 100%)' should be valid in opacity. got disallowed value ""
FAIL 'scale(progress(no-clamp 1%, (10% - 10%), 100%))' as a specified value should serialize as 'scale(calc(0.01))'. assert_not_equals: 'scale(progress(no-clamp 1%, (10% - 10%), 100%))' should be valid in transform. got disallowed value ""
FAIL 'progress(no-clamp 1%, (10% - 10%), 100%)' as a computed value should serialize as '0.01'. assert_equals: 'progress(no-clamp 1%, (10% - 10%), 100%)' and '0.01' should serialize the same in computed values. expected "0.01" but got "1"
FAIL 'scale(progress(no-clamp 1%, (10% - 10%), 100%))' as a computed value should serialize as 'matrix(0.01, 0, 0, 0.01, 0, 0)'. assert_equals: 'scale(progress(no-clamp 1%, (10% - 10%), 100%))' and 'matrix(0.01, 0, 0, 0.01, 0, 0)' should serialize the same in computed values. expected "matrix(0.01, 0, 0, 0.01, 0, 0)" but got "none"
FAIL 'calc(0.5 * progress(no-clamp 100px, 0px, 100px))' as a specified value should serialize as 'calc(0.5)'. assert_not_equals: 'calc(0.5 * progress(no-clamp 100px, 0px, 100px))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(0.5 * progress(no-clamp 100px, 0px, 100px)))' as a specified value should serialize as 'scale(calc(0.5))'. assert_not_equals: 'scale(calc(0.5 * progress(no-clamp 100px, 0px, 100px)))' should be valid in transform. got disallowed value ""
FAIL 'calc(0.5 * progress(no-clamp 100px, 0px, 100px))' as a computed value should serialize as '0.5'. assert_equals: 'calc(0.5 * progress(no-clamp 100px, 0px, 100px))' and '0.5' should serialize the same in computed values. expected "0.5" but got "1"
FAIL 'scale(calc(0.5 * progress(no-clamp 100px, 0px, 100px)))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'. assert_equals: 'scale(calc(0.5 * progress(no-clamp 100px, 0px, 100px)))' and 'matrix(0.5, 0, 0, 0.5, 0, 0)' should serialize the same in computed values. expected "matrix(0.5, 0, 0, 0.5, 0, 0)" but got "none"
FAIL 'calc(0.5 * progress(no-clamp 200px, 0px, 100px))' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'calc(0.5 * progress(no-clamp 200px, 0px, 100px))' should be valid in opacity. got disallowed value ""
FAIL 'calc(0.5 * progress(no-clamp -100px, 0px, 100px))' as a specified value should serialize as 'calc(-0.5)'. assert_not_equals: 'calc(0.5 * progress(no-clamp -100px, 0px, 100px))' should be valid in opacity. got disallowed value ""
FAIL 'progress(no-clamp 2rad, 1rad, 1rad)' as a specified value should serialize as 'calc(infinity)'. assert_not_equals: 'progress(no-clamp 2rad, 1rad, 1rad)' should be valid in opacity. got disallowed value ""
FAIL 'progress(no-clamp 1rad, 1rad, 1rad)' as a specified value should serialize as 'calc(0)'. assert_not_equals: 'progress(no-clamp 1rad, 1rad, 1rad)' should be valid in opacity. got disallowed value ""
FAIL 'progress(no-clamp 0rad, 1rad, 1rad)' as a specified value should serialize as 'calc(-infinity)'. assert_not_equals: 'progress(no-clamp 0rad, 1rad, 1rad)' should be valid in opacity. got disallowed value ""
FAIL 'calc(50px * progress(no-clamp 100px, 0px, 100px))' as a specified value should serialize as 'calc(50px)'. assert_not_equals: 'calc(50px * progress(no-clamp 100px, 0px, 100px))' should be valid in width. got disallowed value ""
FAIL 'calc(1px * progress(no-clamp abs(10%), (10% - 10%), 100% / 10))' as a computed value should serialize as '1px'. assert_equals: 'calc(1px * progress(no-clamp abs(10%), (10% - 10%), 100% / 10))' and '1px' should serialize the same in computed values. expected "1px" but got "784px"
